backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.33.0
pandas==2.1.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.8.3
sqlparse==0.5.0